+++ experimental/tiledb/common/dag/utility/range_join.h
@@ -73,7 +73,7 @@
 #include <type_traits>
 #include <utility>
 #include <vector>
-#include "tiledb/common/arrow_proxy.hpp"
+#include "tiledb/common/util/arrow_proxy.h"
 
 #include "external/include/span/span.hpp"
