+++ tiledb/common/util/CMakeLists.txt
@@ -1,9 +1,9 @@
 #
-# tiledb/common/test/CMakeLists.txt
+# tiledb/common/thread_pool/CMakeLists.txt
 #
 # The MIT License
 #
-# Copyright (c) 2022 TileDB, Inc.
+# Copyright (c) 2021 TileDB, Inc.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -24,25 +24,18 @@
 # THE SOFTWARE.
 #
 
-include(unit_test)
+include(common NO_POLICY_SCOPE)
+include(object_library)
 
-commence(unit_test experimental)
-    this_target_object_libraries(baseline)
-    this_target_sources(
-        main.cc
-        $<IF:$<BOOL:${TILEDB_EXPERIMENTAL_FEATURES}>,
-          unit_is_experimental.cc,
-          unit_is_not_experimental.cc>
-        )
-conclude(unit_test)
+# list(APPEND SOURCES)
+# gather_sources(${SOURCES})
 
-commence(unit_test memory_tracker)
-    this_target_sources(main.cc unit_memory_tracker.cc unit_memory_tracker_types.cc)
-    this_target_object_libraries(baseline)
-conclude(unit_test)
+#
+# There is no object library for this target, as it is a header-only library
+#
+# commence(object_library util)
+#    this_target_sources(${SOURCES})
+#    this_target_object_libraries(baseline)
+# conclude(object_library)
 
-commence(unit_test common_utils)
-    this_target_sources(main.cc ../util/test/unit_alt_var_length_view.cc ../util/test/unit_iterator_facade.cc ../util/test/unit_permutation_sort.cc ../util/test/unit_permutation_view.cc ../util/test/unit_proxy_sort.cc ../util/test/unit_sort_zip.cc ../util/test/unit_var_length_util.cc ../util/test/unit_var_length_view.cc ../util/test/unit_zip_view.cc)
-    this_target_object_libraries(baseline)
-conclude(unit_test)
-
+add_test_subdirectory()
