+++ tiledb/common/util/test/CMakeLists.txt
@@ -1,9 +1,9 @@
 #
-# tiledb/common/test/CMakeLists.txt
+# tiledb/common/util/CMakeLists.txt
 #
 # The MIT License
 #
-# Copyright (c) 2022 TileDB, Inc.
+# Copyright (c) 2021 TileDB, Inc.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -26,23 +26,16 @@
 
 include(unit_test)
 
-commence(unit_test experimental)
-    this_target_object_libraries(baseline)
-    this_target_sources(
-        main.cc
-        $<IF:$<BOOL:${TILEDB_EXPERIMENTAL_FEATURES}>,
-          unit_is_experimental.cc,
-          unit_is_not_experimental.cc>
-        )
+commence(unit_test util)
+this_target_sources(
+        unit_alt_var_length_view.cc
+        unit_iterator_facade.cc
+        unit_permutation_sort.cc
+        unit_permutation_view.cc
+        unit_proxy_sort.cc
+        unit_sort_zip.cc
+        unit_var_length_util.cc
+        unit_var_length_view.cc
+        unit_zip_view.cc
+)
 conclude(unit_test)
-
-commence(unit_test memory_tracker)
-    this_target_sources(main.cc unit_memory_tracker.cc unit_memory_tracker_types.cc)
-    this_target_object_libraries(baseline)
-conclude(unit_test)
-
-commence(unit_test common_utils)
-    this_target_sources(main.cc ../util/test/unit_alt_var_length_view.cc ../util/test/unit_iterator_facade.cc ../util/test/unit_permutation_sort.cc ../util/test/unit_permutation_view.cc ../util/test/unit_proxy_sort.cc ../util/test/unit_sort_zip.cc ../util/test/unit_var_length_util.cc ../util/test/unit_var_length_view.cc ../util/test/unit_zip_view.cc)
-    this_target_object_libraries(baseline)
-conclude(unit_test)
-
