+++ tiledb/common/util/test/unit_permutation_sort.cc
@@ -36,11 +36,11 @@
 #include <random>
 #include <ranges>
 #include <vector>
-#include "../alt_var_length_view.h"
-#include "../permutation_view.h"
-#include "../proxy_sort.h"
-#include "../var_length_view.h"
-#include "../zip_view.h"
+#include "tiledb/common/util/alt_var_length_view.h"
+#include "tiledb/common/util/permutation_view.h"
+#include "tiledb/common/util/proxy_sort.h"
+#include "tiledb/common/util/var_length_view.h"
+#include "tiledb/common/util/zip_view.h"
 
 TEST_CASE("permutation_sort: Null test", "[permutation_sort][null_test]") {
   REQUIRE(true);
