+++ tiledb/common/util/test/unit_sort_zip.cc
@@ -33,8 +33,8 @@
 #include <algorithm>
 #include <catch2/catch_all.hpp>
 #include <vector>
-#include "../alt_var_length_view.h"
-#include "../zip_view.h"
+#include "tiledb/common/util/alt_var_length_view.h"
+#include "tiledb/common/util/zip_view.h"
 
 TEST_CASE("sort_zip: Null test", "[zip_view][null_test]") {
   REQUIRE(true);
